+++ src/GDD.Core/GddVersion.cs
@@ -2,5 +2,5 @@ namespace GDD;
 
 public static class GddVersion
 {
-    public const string Current = "1.4.3";
+    public const string Current = "1.4.4";
 }